Expose underlying tap-spec stream
